requirements.txt
simplejson==3.20.1
beautifulsoup4==4.13.4
html5lib==1.1
regex==2024.11.6
redis==6.1.0
python-dotenv==1.1.0
requests==2.32.3
textblob==0.19.0
lxml==5.4.0
boto3==1.38.37
numpy==2.2.6
bitarray==3.7.2
playwright==1.56.0
## The following requirements were added by pip freeze:
async-timeout==5.0.1
botocore==1.38.37
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
greenlet==3.2.4
idna==3.10
jellyfish==1.2.1
jmespath==1.0.1
joblib==1.5.0
metadata_parser==0.13.0
nltk==3.9.1
pyee==13.0.0
python-dateutil==2.9.0.post0
requests-toolbelt==1.0.0
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
lingua-language-detector==2.1.1
